% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.r
\name{culvert_area}
\alias{culvert_area}
\alias{culvert_wetted_perimeter}
\alias{culvert_hydraulic_radius}
\title{Culvert Hydraulic Geometry}
\usage{
culvert_area(h, d = 5, leaf = c("square", "circular"))

culvert_wetted_perimeter(h, d = 5, leaf = c("square", "circular"))

culvert_hydraulic_radius(h, d, leaf = c("square", "circular"))
}
\arguments{
\item{h}{Slide gate height, ft.}

\item{d}{Culvert diameter, ft. Default is 5 ft.}

\item{leaf}{The leaf type, either \code{"circular"} or \code{"square"}.}
}
\value{
Culvert area, ft^2.

Wetted perimeter, ft.

Hydraulic radius, ft.
}
\description{
Functions for calculating culvert flow area, wetted perimeter, and
hydraulic radius.
}
