% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{gate-loss}
\alias{gate-loss}
\alias{usda_ks}
\alias{brater_ks}
\title{Slide Gate Head Loss Tables}
\usage{
usda_ks()

brater_ks()
}
\value{
A two-column \code{data.frame}. First column is
the gate opening height to culvert diameter ratio \eqn{h / d},
Second column is the coefficient value.
}
\description{
Tabulated coefficients for selected gate opening fractions.
}
\details{
\code{\link[=usda_ks]{usda_ks()}} uses \code{ks} values from the USDA W.E.S. Hyd.
Chart 330-1, "Gate headloss coefficient (Kg) full pipe flow"
(USDA, 1969, p. B-25).

\code{\link[=brater_ks]{brater_ks()}} uses \code{ks} values from Brater and King
"Handbook of hydraulics for the solution of hydraulic engineering
problems" (1976).
}
