% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_entrance}
\alias{loss_entrance}
\title{Entrance Head Loss}
\usage{
loss_entrance(v, ke = 0.9, g = 32.2)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{ke}{Entrance loss coefficient. Default is 0.9 (USDOT 2005).}

\item{g}{Gravitational acceleration. Default value is 32.2 ft/s^2.}
}
\value{
The entrance head loss, ft.
}
\description{
Culvert entrance head loss
}
