% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_flap}
\alias{loss_flap}
\title{Flap Gate Head Loss}
\usage{
loss_flap(v, d = 5, kf = 1, g = 32.2)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{d}{Pipe diameter, ft. Default value is 5 ft.}

\item{kf}{Flap gate weight correction factor.}

\item{g}{Gravitational acceleration. Default value is 32.2 ft/s^2.}
}
\value{
Flap gate head loss, ft.
}
\description{
Head loss at flap gate.
}
