% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_obstruction}
\alias{loss_obstruction}
\title{Head Loss From Obstruction}
\usage{
loss_obstruction(v, h, d = 5, ko.table = king_ko(), g = 32.2, ...)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{h}{Slide gate opening, ft.}

\item{d}{Pipe diameter, ft. Default value is 5 ft.}

\item{ko.table}{Obstruction head loss table. Used to determine the
value of \code{ko} as a function of the slide gate opening height.
Default uses the built-in \code{\link[=usda_ks]{usda_ks()}} lookup table.}

\item{g}{Gravitational acceleration. Default value is 32.2 ft/s^2.}

\item{...}{Additional arguments to \code{\link[=culvert_area]{culvert_area()}}.}
}
\description{
Head loss resulting from obstruction inside pipe, e.g., a
partially-closed valve. Provides an alternative approach to
Computing slide gate loss.
}
\details{
This approach may be preferable to \code{\link[=loss_slide]{loss_slide()}} because
this equation assumes the obstruction is inside the pipe, whereas
the slide gate loss equation assumes the slide gate is at the
pipe entrance. Because the Roaring River Distribution System
culvert slide gates are inside the pipes, the obstruction loss
equation may be a better approximation of the loss.
}
