% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_outlet}
\alias{loss_outlet}
\title{Outlet Head Loss}
\usage{
loss_outlet(v, vd = 0, alpha = 1, g = 32.2)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{vd}{Downstream velocity, ft/s. Default value is 0.}

\item{alpha}{Outlet coefficient. Default value is 1.}

\item{g}{Gravitational acceleration. Default value is 32.2 ft/s^2.}
}
\value{
Outlet head loss, ft.
}
\description{
Culvert outlet head loss.
}
