% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_pipe}
\alias{loss_pipe}
\title{Pipe Friction Head Loss}
\usage{
loss_pipe(v, d = 5, l = 98.5, ku = 29.164, n = 0.025, g = 32.2)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{d}{Pipe diameter, ft. Default value is 5 ft.}

\item{l}{Pipe length, ft. Default value is 98.5 ft.}

\item{ku}{Loss coefficient. Default value is 29.}

\item{n}{Mannings roughness coeffient. Default value is 0.1.}

\item{g}{Gravitational acceleration. Default value is 32.2 ft/s^2.}
}
\value{
Pipe friction head loss, ft.
}
\description{
Friction head loss from pipe geometry and surface roughness.
}
