% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_slide}
\alias{loss_slide}
\title{Slide Gate Head Loss}
\usage{
loss_slide(v, h, d = 5, ks.table = usda_ks(), g = 32.2)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{h}{Slide gate opening, ft.}

\item{d}{Pipe diameter, ft. Default value is 5 ft.}

\item{ks.table}{Slide gate head loss table. Used to determine the
value of \code{ks} as a function of the ratio of the slide gate
opening height to pipe diameter. Default uses the built-in
\code{\link[=usda_ks]{usda_ks()}} lookup table.}

\item{g}{Gravitational acceleration. Default value is 32.2 ft/s^2.}
}
\value{
The entrance head loss, ft.
}
\description{
Culvert slide gate head loss
}
\details{
Coefficient values are based on a linear interpolation
of values specified in \code{ks.table}. The nearest (i.e., minimum or
maximum) coefficient value is used for gate opening values outside
the range specified in the table.
}
\seealso{
\code{\link[=usda_ks]{usda_ks()}} \code{\link[=brater_ks]{brater_ks()}}
}
