% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss.r
\name{loss_total}
\alias{loss_total}
\alias{loss_total_hist}
\alias{loss_total_alt}
\title{Total Head Loss}
\usage{
loss_total(...)

loss_total_hist(...)

loss_total_alt(...)
}
\arguments{
\item{...}{Arguments to the component loss functions, i.e.,
\code{\link[=loss_entrance]{loss_entrance()}}, \code{\link[=loss_slide]{loss_slide()}}, \code{\link[=loss_obstruction]{loss_obstruction()}},
\code{\link[=loss_pipe]{loss_pipe()}}, \code{\link[=loss_outlet]{loss_outlet()}}, and \code{\link[=loss_flap]{loss_flap()}}.}
}
\value{
Total head loss, ft.
}
\description{
Total head loss, i.e., combined entrance, slide gate/obstruction,
pipe friction, outlet, and junction chamber losses.
}
\details{
the total head loss formula is
\deqn{
    L_{total} = L_{entrance} + L_{pipe friction} + L_{slide gate} +
      L_{outlet} + L_{flap gate}
  }
where \eqn{L_{outlet}} accounts for  junction chamber connection
and \eqn{L_{flap gate}} accounts for the culvert outlet.
Note that the total head loss equation was originally designed
to use \eqn{ks} values from \code{usda_ks()} for computing slide gate
losses.

the historical total head loss formula is
\deqn{
    L_{total} = L_{entrance} + L_{pipe friction} + L_{slide gate} +
    2L_{outlet}
  }
where the same outlet loss equation is used for both the junction
chamber connection and the culvert outlet.
Note that the historical total head loss equation was originally
designed to use \eqn{ks} values from \code{brater_ks()} for computing
slide gate losses. Also, the outlet loss term was originally
designed using an alternative parameterization of friction loss
which is approximately equivalent to an \code{n} value of 0.026.

the alternate total head loss formula is
\deqn{
    L_{total} = L_{entrance} + L_{pipe friction} + L_{obstruction} +
      L_{outlet} + L_{flap gate}
  }
where the obstruction loss term replaces the slide gate loss term.
}
