% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.r
\name{obstruction-loss}
\alias{obstruction-loss}
\alias{king_ko}
\title{Culvert Obstruction Loss Coefficients}
\usage{
king_ko()
}
\value{
A two-column \code{data.frame}. First column is
the total culvert area to unobstructed area ratio \eqn{A / A_o},
Second column is the coefficient value.
}
\description{
Tabulated coefficients for selected ratios of total culvert area to
unobstructed area.
}
\details{
\code{\link[=king_ko]{king_ko()}} uses \code{ko} values from the National Hydraulic ###
}
