% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.r
\name{root_velocity}
\alias{root_velocity}
\alias{root_velocity_hist}
\alias{root_velocity_alt}
\title{Solve for Culvert Velocity/Flow}
\usage{
root_velocity(v, head.loss, ...)

root_velocity_hist(v, head.loss, ...)

root_velocity_alt(v, head.loss, ...)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{head.loss}{Total head loss, ft.}

\item{...}{Arguments to the component loss functions, i.e.,
\code{\link[=loss_entrance]{loss_entrance()}}, \code{\link[=loss_slide]{loss_slide()}}, \code{\link[=loss_obstruction]{loss_obstruction()}},
\code{\link[=loss_pipe]{loss_pipe()}}, \code{\link[=loss_outlet]{loss_outlet()}}, and \code{\link[=loss_flap]{loss_flap()}}.}
}
\description{
Root equation for solving culvert velocity or flow given a known
head loss. This function simply returns the difference of the
provided head loss and the evaluated result of \code{\link{loss_total}}
or \code{\link{loss_total_alt}}. These functions are designed to be used
in conjunction with \code{\link[stats:uniroot]{stats::uniroot()}}.
}
