% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.r
\name{velocity_to_flow}
\alias{velocity_to_flow}
\alias{velocity_from_flow}
\title{Velocity/Flow Conversion}
\usage{
velocity_to_flow(v, d = 5)

velocity_from_flow(q, d = 5)
}
\arguments{
\item{v}{Fluid velocity, ft/s}

\item{d}{Culvert diameter, ft. Default is 5 ft.}

\item{q}{Fluid flow, ft^3/s.}
}
\value{
Fluid flow, ft^3/s.

Fluid velocity, ft/s.
}
\description{
Simple helper function to convert between flow and velocity.
These functions assume full flow through a circular pipe/culvert.
}
\details{
Flow is calculated as the product of velocity and culvert
cross-sectional area.
}
